pyproject.toml
[project]
name = "fleetcare-data-harvest"
version = "1.1.0"
description = "A basic web service to act as a webhook for an Azure EventGrid subscription (BlobCreated events)"
authors = [{ name = "Ashley Felton", email = "ashley.felton@dbca.wa.gov.au" }]
license = "Apache-2.0"
requires-python = ">=3.13,<4.0"

dependencies = [
  "quart==0.20.0",
  "hypercorn==0.17.3",
  "python-dotenv==1.1.0",
  "quart-db[postgresql]>=0.9.0",
  "azure-storage-blob==12.25.1",
]

[dependency-groups]
dev = ["ipdb>=0.13.13", "ipython>=8.31.0", "pre-commit>=4.2.0"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
line-length = 140

[tool.ruff.lint]
ignore = [
  "E501", # Line too long
  "E722", # Bare except
]
